# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Education Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: Chat operations
  - name: knowledge
    description: Knowledge base operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /chat/sessions:
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{sessionId}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get AI response
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /knowledge:
    get:
      operationId: getKnowledgeBase
      tags: [knowledge]
      summary: Get all knowledge base entries
      responses:
        "200":
          description: Knowledge base entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeEntry"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ChatSession:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - createdAt
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        confidence:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - role
        - content
        - createdAt
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    KnowledgeEntry:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
        category:
          type: string
          nullable: true
      required:
        - id
        - question
        - answer