# lib/api-zod/src/generated/types/chatSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Education Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSession {
  id: string;
  createdAt: Date;
}